Offloading some of the functions in the router (index.js) to files in the contoller folder.
